cluster_image/pyproject.toml
[project]
name = "cluster-image"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.5",
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "plotly>=6.2.0",
    "scikit-learn>=1.7.1",
    "seaborn>=0.13.2",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.55.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.0",
    "jupyter>=1.1.1",
    "pytest>=8.4.1",
]
